feat: Implement advanced message and campaign management
