+++ app/src/main/java/com/quansoft/smsgateway/data/local/SmsMessageEntity.kt
@@ -1,4 +1,42 @@
 package com.quansoft.smsgateway.data.local
 
-class SmsMessageEntity {
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.Index
+import androidx.room.PrimaryKey
+import com.quansoft.smsgateway.domain.model.Message
+
+@Entity(
+    tableName = "sms_messages",
+    foreignKeys = [
+        ForeignKey(
+            entity = CampaignEntity::class,
+            parentColumns = ["id"],
+            childColumns = ["bulkId"],
+            onDelete = ForeignKey.CASCADE
+        )
+    ],
+    indices = [Index(value = ["bulkId"])]
+)
+data class SmsMessageEntity(
+    @PrimaryKey val id: String,
+    val recipient: String,
+    val content: String,
+    var status: String,
+    val timestamp: Long,
+    val bulkId: String? = null
+) {
+
+}
+
+// Mapper function to convert Entity to Domain model
+fun SmsMessageEntity.toDomain(): Message {
+    return Message(
+        id = this.id,
+        recipient = this.recipient,
+        content = this.content,
+        status = this.status,
+        timestamp = this.timestamp,
+        bulkId = this.bulkId
+    )
 }
