+++ app/src/main/java/com/quansoft/smsgateway/domain/model/Campaign.kt
@@ -1,4 +1,8 @@
+
 package com.quansoft.smsgateway.domain.model
 
-class Campaign {
-}
+data class Campaign(
+    val id: String,
+    val name: String,
+    val timestamp: Long
+)
