+++ app/src/main/java/com/quansoft/smsgateway/domain/model/Message.kt
@@ -1,4 +1,10 @@
 package com.quansoft.smsgateway.domain.model
 
-class Message {
-}
+data class Message(
+    val id: String,
+    val recipient: String,
+    val content: String,
+    val status: String,
+    val timestamp: Long,
+    val bulkId: String?
+)
