+++ app/src/main/java/com/quansoft/smsgateway/domain/model/MessageWithDetails.kt
@@ -1,4 +1,11 @@
 package com.quansoft.smsgateway.domain.model
 
-class MessageWithDetails {
-}
+/**
+ * A domain model that combines a Message with its related details
+ * like contact and campaign names for UI presentation.
+ */
+data class MessageWithDetails(
+    val message: Message,
+    val contactName: String?,
+    val campaignName: String?
+)
