+++ app/src/main/java/com/quansoft/smsgateway/domain/repository/CampaignRepository.kt
@@ -1,4 +1,9 @@
 package com.quansoft.smsgateway.domain.repository
+import com.quansoft.smsgateway.domain.model.Campaign
+import kotlinx.coroutines.flow.Flow
 
-class CampaignRepository {
+interface CampaignRepository {
+    fun getAllCampaigns(): Flow<List<Campaign>>
+    suspend fun deleteCampaignAndMessages(campaign: Campaign)
+    suspend fun delete(campaign: Campaign)
 }
