+++ app/src/main/java/com/quansoft/smsgateway/domain/repository/MessageRepository.kt
@@ -1,4 +1,11 @@
 package com.quansoft.smsgateway.domain.repository
 
-class MessageRepository {
+import com.quansoft.smsgateway.domain.model.Message
+import kotlinx.coroutines.flow.Flow
+
+interface MessageRepository {
+    fun getAllMessages(): Flow<List<Message>>
+    suspend fun deleteMessage(message: Message)
+    suspend fun delete(message: Message)
+    suspend fun deleteMessagesByCampaignId(id: String)
 }
