+++ app/src/main/java/com/quansoft/smsgateway/domain/repository/SettingsRepository.kt
@@ -1,4 +1,11 @@
 package com.quansoft.smsgateway.domain.repository
 
-class SettingsRepository {
+import kotlinx.coroutines.flow.Flow
+
+interface SettingsRepository {
+    fun getServerPort(): Flow<Int>
+    suspend fun setServerPort(port: Int)
+    fun getAuthToken(): Flow<String>
+    suspend fun regenerateAuthToken(): String
+    suspend fun generateAndStoreInitialToken(): String
 }
