+++ app/src/main/java/com/quansoft/smsgateway/service/BootReceiver.kt
@@ -8,7 +8,7 @@ import android.os.Build
 class BootReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
         if (intent.action == Intent.ACTION_BOOT_COMPLETED) {
-            val serviceIntent = Intent(context, GatewayService::class.java)
+            val serviceIntent = Intent(context, SMSService::class.java)
 
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 context.startForegroundService(serviceIntent)
