+++ app/src/main/java/com/quansoft/smsgateway/service/SmsStatusReceiver.kt
@@ -4,7 +4,7 @@ import android.app.Activity
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import com.quansoft.smsgateway.data.AppDatabase
+import com.quansoft.smsgateway.data.local.AppDatabase
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
