+++ app/src/main/java/com/quansoft/smsgateway/ui/info/InfoScreen.kt
@@ -33,6 +33,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.quansoft.smsgateway.R
+import com.quansoft.smsgateway.ui.widgets.InfoCard
 import com.quansoft.smsgateway.util.AppUtils
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -115,32 +116,3 @@ fun AppHeader(version: String) {
     }
 }
 
-@Composable
-fun InfoCard(icon: ImageVector, title: String, content: String) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Icon(
-                    imageVector = icon,
-                    contentDescription = null,
-                    modifier = Modifier.size(24.dp),
-                    tint = MaterialTheme.colorScheme.primary
-                )
-                Spacer(modifier = Modifier.width(16.dp))
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.titleLarge,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = content,
-                style = MaterialTheme.typography.bodyLarge
-            )
-        }
-    }
-}
