+++ app/src/main/java/com/quansoft/smsgateway/ui/widgets/MessageList.kt
@@ -1,4 +1,103 @@
 package com.quansoft.smsgateway.ui.widgets
 
-class MessageList {
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SwipeToDismissBox
+import androidx.compose.material3.SwipeToDismissBoxValue
+import androidx.compose.material3.Text
+import androidx.compose.material3.rememberSwipeToDismissBoxState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.quansoft.smsgateway.domain.model.MessageWithDetails
+import kotlinx.coroutines.delay
+
+
+@Composable
+fun MessageList(messages: List<MessageWithDetails>, onDelete: (MessageWithDetails) -> Unit) {
+    if (messages.isEmpty()) {
+        Box(modifier = Modifier.fillMaxSize().padding(16.dp), contentAlignment = Alignment.Center) {
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                Icon(Icons.Default.Search, contentDescription = "No Messages", modifier = Modifier.size(64.dp), tint = Color.Gray)
+                Spacer(Modifier.height(16.dp))
+                Text("No messages found", style = MaterialTheme.typography.headlineSmall, textAlign = TextAlign.Center)
+                Text("Try adjusting your search or filter criteria.", style = MaterialTheme.typography.bodyMedium, textAlign = TextAlign.Center, color = Color.Gray)
+            }
+        }
+    } else {
+        LazyColumn(
+            verticalArrangement = Arrangement.spacedBy(8.dp),
+            contentPadding = PaddingValues(vertical = 8.dp, horizontal = 16.dp)
+        ) {
+            items(messages, key = { it.message.id }) { item ->
+                var isDismissed by remember { mutableStateOf(false) }
+                val dismissState = rememberSwipeToDismissBoxState(
+                    confirmValueChange = {
+                        if (it == SwipeToDismissBoxValue.EndToStart) {
+                            isDismissed = true
+                            true
+                        } else false
+                    }
+                )
+
+                LaunchedEffect(isDismissed) {
+                    if (isDismissed) {
+                        delay(300) // Wait for animation
+                        onDelete(item)
+                    }
+                }
+
+                SwipeToDismissBox(
+                    state = dismissState,
+                    backgroundContent  = {
+                        val color = when (dismissState.targetValue) {
+                            SwipeToDismissBoxValue.EndToStart -> MaterialTheme.colorScheme.errorContainer
+                            else -> Color.Transparent
+                        }
+                        val icon = Icons.Default.Delete
+                        Box(
+                            modifier = Modifier
+                                .fillMaxSize()
+                                .clip(CardDefaults.shape)
+                                .background(color)
+                                .padding(16.dp),
+                            contentAlignment = Alignment.CenterEnd
+                        ) {
+                            Icon(
+                                imageVector = icon,
+                                contentDescription = "Delete Icon",
+                                tint = MaterialTheme.colorScheme.onErrorContainer
+                            )
+                        }
+                    },
+                    content = { MessageItem(item,onDelete = { onDelete(item) }) }
+                )
+            }
+        }
+    }
 }
